# Conflicts:
#	src/devices/Arduino/ArduinoBoard.cs
#	src/devices/Axp192/Axp192.csproj
#	src/devices/Axp192/M5ToughPowerControl.cs
#	src/devices/Axp192/PowerControlData.cs
#	src/devices/Common/Iot/Device/Graphics/LowLevelX11Window.cs
#	src/devices/Common/Iot/Device/Graphics/MouseClickSimulatorUInput.uinput.cs
#	src/devices/Common/Iot/Device/Graphics/MouseClickSimulatorX11.Linux.cs
#	src/devices/Common/Iot/Device/Graphics/ScreenCapture.Linux.cs
#	src/devices/Common/Iot/Device/Graphics/ScreenCapture.Windows.cs
#	src/devices/Common/Iot/Device/Graphics/ScreenCapture.cs
#	src/devices/Common/Iot/Device/Graphics/WindowsTouchSimulator.Windows.cs
#	src/devices/Ili9341/category.txt
#	src/devices/Ili934x/Chsc6440.cs
#	src/devices/Ili934x/DragEventArgs.cs
#	src/devices/Ili934x/Ili9341.cs
#	src/devices/Ili934x/Ili9341Imaging.cs
#	src/devices/Ili934x/Ili9342.cs
#	src/devices/Ili934x/Ili934x.sln
#	src/devices/Ili934x/README.md
#	src/devices/Ili934x/Rgb565.cs
#	src/devices/Ili934x/category.txt
#	src/devices/Ili934x/samples/Ili934x.Samples.csproj
#	src/devices/Ili934x/samples/NmeaValueDataSet.cs
#	src/devices/Ili934x/samples/Program.cs
#	src/devices/Ili934x/samples/RemoteControl.cs
#	src/devices/Is31fl3730/category.txt

+++ src/devices/Ili934x/samples/NmeaValueDataSet.cs
@@ -64,7 +64,7 @@ namespace Iot.Device.Ili934x.Samples
                     return true;
                 }
 
-                bool ret = Math.Abs(newValue.Value - _lastValue.Value) > tolerance;
+                bool ret = Math.Abs((double)newValue.Value - (double)_lastValue.Value) > tolerance;
                 _lastValue = newValue;
 
                 return ret;
